pyproject.toml
[project]
name = "trigger-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "viam-sdk>=0.47.0",
]

[dependency-groups]
lint = [
    "ruff>=0.12.0",
]
